Add container background
