+++ ios/Widget/Backport.swift
@@ -1,0 +1,26 @@
+import WidgetKit
+import SwiftUI
+
+extension WidgetConfiguration {
+  @_disfavoredOverload
+  func contentMarginsDisabled() -> some WidgetConfiguration {
+    if #available(iOSApplicationExtension 17.0, *) {
+      return self.contentMarginsDisabled()
+    } else {
+      return self
+    }
+  }
+}
+
+extension View {
+  func widgetBackground(_ backgroundView: some View) -> some View {
+    if #available(iOSApplicationExtension 17.0, *) {
+      return containerBackground(for: .widget) {
+        backgroundView
+      }
+    } else {
+      return background(backgroundView)
+    }
+  }
+}
+
